Renamed serial reader files
